+++ CHANGELOG.md
@@ -3,8 +3,12 @@
 ### Breaking changes
 
 * Renamed `User` to `SyncUser`, `Credentials` to `SyncCredentials` and `Session` to `SyncSession` to align names with Cocoa.
-* Added `Logger.log()` to replaces the mutiple methods in the `Logger` interface.
-* Removed `SyncManager.setLogLevel()` and `Util.setDebugLevel()`. Added `RealmLog.setLevel()` instead.
+
+### Deprecated
+
+* `Logger`. Use `RealmLogger` instead.
+* `AndroidLogger`. The logger for Android is implemented in the native code instead.
+* `SyncManager.setLogLevel()`. Use `RealmLog.setLevel()` instead.
 
 ### Bug fixes